Address review 2, issue 1
Q: Why not model a moving landmark as a sequence of "dynamic"
variables?

Provide justification: Because scalar mobility variables are enough
for the job. Do not need to model the kinematics, if that is what
"dynamic" means.

It looks like there is some confusion caused by nested subscripts
in w_{j_k}. "z_i" can not really depend on "w_i".

Provide more specific explanation of j_k in w_{j_k}.