Address review 2, issue 2
Q: why w_{j_k} appears inside the exponential in eq. 4?
Q: how to derive eq. (4)?

Eq. (1)-(4) are really all definitions. w_{j_k} being inside the
exponential is part of the definition in eq. (4). Unfortunately
w_{j_k} is not explicitly defined as a conditional distribution.
We have given very clear meanings for w_{j_k}: a scaling factor
of the covariance. We did not find an elegant form to formulate
w_{j_k} as a prior of l_{j_k} therefore it is defined implicitly.

Change wordings to stress these are all definitions.